requirements.txt
flask==3.1.2
flask-cors==6.0.1
PyJWT==2.8.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
